模型/requirements.txt
flask==3.0.3
flask-cors==5.0.0
h5py==3.13.0
imageio==2.37.0
joblib==1.4.2
matplotlib==3.10.0
numpy==1.26.4
opencv-python==4.10.0.84
pandas==2.2.3
Pillow==10.4.0
python-dotenv==1.0.1
scikit-image==0.25.1
scikit-learn==1.6.1
scipy==1.15.2
tensorboardX==2.6.2.2
timm==1.0.15
torch==2.6.0
torchvision==0.21.0
